Renam to wallet.js & refactor
