+++ .gitignore
@@ -4,3 +4,5 @@ bower_components/
 
 build/
 client/dist/bundle.js
+client/config/config.js
+client/dist/bundle.js
