+++ _.gitignore
@@ -21,6 +21,7 @@ coverage
 # Compiled Dirs (http://nodejs.org/api/addons.html)
 build/
 client/dist/bundle.js
+config/config.js
 
 # Dependency directorys
 # Deployed apps should consider commenting these lines out:
